--- src/components/ChartDateRange/index.ts
@@ -1,2 +1,0 @@
-export { default as ChartDateRange } from './ChartDateRange.vue'
-export * from './types'
