+++ src/components/ChartZoom/index.ts
@@ -1,0 +1,2 @@
+export { default as ChartZoom } from './ChartZoom.vue'
+export * from './types'
