+++ src/components/index.ts
@@ -3,7 +3,7 @@ export { default as DivergingBarChart } from './DivergingBarChart.vue'
 export { default as Timeline } from './Timeline.vue'
 export { default as ScatterPlot } from './ScatterPlot.vue'
 export * from './ChartCursor'
-export * from './ChartDateRange'
+export * from './ChartZoom'
 export * from './ChartLabels'
 export * from './ChartPlayHead'
 export * from './ChartSelection'
